Refactor the cli modules structure
