+++ crates/humanode-peer/src/cli/mod.rs
@@ -1,0 +1,11 @@
+//! Command line interface.
+
+mod root;
+mod run;
+mod runner;
+mod subcommand;
+
+pub use root::*;
+pub use run::*;
+pub use runner::*;
+pub use subcommand::*;
